Refactor: Change parameter, making header struct, function callback
Change parameters to use reference than pointer to prevent memory copy.

Make Get,SetHeader functions to check packet size and record it in packet.
We can get packet's size when entire data is serialized.
So, setting header must locate in the last of the process.

Calling functions with with map<PacketType, function pointer> is much slower than lambda function call.
Because of this, changed the way to execute function into switch.